pyproject.toml
[project]
name = "deforisk-jupyter-nb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.10"
dependencies = [
    "cudf-cu12>=25.8.0",
    "cuml-cu12>=25.8.0",
    "dask-ml>=2025.1.0",
    "fiona>=1.10.1",
    "forestatrisk>=1.3.2",
    "gdal[numpy]==3.8.4",
    "geedim>=2.0.0",
    "geemap>=0.36.2",
    "geopandas>=1.1.1",
    "numpy==1.26.4",
    "odc-geo>=0.4.10",
    "pycrs>=1.0.2",
    "rasterio>=1.4.3",
    "rioxarray>=0.19.0",
    "riskmapjnr>=1.3.2",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
    "xarray-spatial>=0.4.0",
]

[project.license]
text = "MIT"

[[tool.uv.index]]
name = "gdal-wheels"
url = "https://gitlab.com/api/v4/projects/61637378/packages/pypi/simple"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.ruff]
select = ["E", "F", "W", "I", "D", "RUF"]
exclude = [".venv"]
ignore = ["D200","D212"]
[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"*/__init__.py" = ["F403"] # unable to detect undefined names | hide internal structure
"setup.py" = ["D100"] # nothing to see there
